// src/types.d.ts
export interface IRawPacket {
  connectionId: string
  connection: TCPConnection
  data: Buffer
  localPort: number | undefined
  remoteAddress: string | undefined
  timestamp: number
}
// ...
export interface ISessionRecord {
  skey: string
  sessionkey: string
}